feat(welcomechannel): bikin sub command group biar rapih
