+++ src/commands/interactions/welcomechannel/setup.ts
@@ -1,4 +1,4 @@
-import { AutoLoad, Command, Declare, IgnoreCommand } from "seyfert";
+import { AutoLoad, Command, Declare, Groups, IgnoreCommand } from "seyfert";
 
 @Declare({
   name: "setup-channel",
@@ -7,4 +7,9 @@ import { AutoLoad, Command, Declare, IgnoreCommand } from "seyfert";
   ignore: IgnoreCommand.Message,
 })
 @AutoLoad()
+@Groups({
+  welcome: {
+    defaultDescription: "set up channel for welcome and out member",
+  },
+})
 export default class SetupChanmel extends Command {}
